+++ plugins/hu.eltesoft.modelexecution.ide.debug/src/hu/eltesoft/modelexecution/ide/debug/model/XUMLRTDebugTarget.java
@@ -187,7 +187,7 @@ public class XUMLRTDebugTarget extends DelegatingDebugTarget {
 	public void markThreadAsSuspended(EObject modelElement) {
 		isSuspended = true;
 
-		Pair<String, Integer> actualSMInstance = vmBrowser.getActualSMInstance();
+		Pair<String, Long> actualSMInstance = vmBrowser.getActualSMInstance();
 		List<StateMachineInstance> smInstances = getSmInstances();
 		for (StateMachineInstance smInstance : smInstances) {
 			StateMachineStackFrame stackFrame;
@@ -276,7 +276,7 @@ public class XUMLRTDebugTarget extends DelegatingDebugTarget {
 	 *         null if there isn't one.
 	 */
 	public StateMachineInstance getSuspendedSMInstance() {
-		Pair<String, Integer> actualSMInstance = vmBrowser.getActualSMInstance();
+		Pair<String, Long> actualSMInstance = vmBrowser.getActualSMInstance();
 		for (StateMachineInstance instance : getSmInstances()) {
 			if (actualSMInstance.equals(new Pair<>(instance.getClassId(), instance.getInstanceId()))) {
 				return instance;
